IOT-44: Authorization on Datatarget CRUD

+++ src/controllers/admin-controller/data-target.controller.ts
@@ -9,8 +9,17 @@ import {
     Delete,
     NotFoundException,
     Query,
+    UseGuards,
+    Req,
 } from "@nestjs/common";
-import { ApiTags, ApiOperation, ApiBadRequestResponse } from "@nestjs/swagger";
+import {
+    ApiTags,
+    ApiOperation,
+    ApiBadRequestResponse,
+    ApiBearerAuth,
+    ApiForbiddenResponse,
+    ApiUnauthorizedResponse,
+} from "@nestjs/swagger";
 import { ListAllDataTargetsReponseDto } from "@dto/list-all-data-targets-response.dto";
 import { CreateDataTargetDto } from "@dto/create-data-target.dto";
 import { DataTarget } from "@entities/data-target.entity";
@@ -19,27 +28,65 @@ import { DeleteResponseDto } from "@dto/delete-application-response.dto";
 import { ErrorCodes } from "@enum/error-codes.enum";
 import { DataTargetService } from "@services/data-target.service";
 import { ListAllDataTargetsDto } from "@dto/list-all-data-targets.dto";
+import { JwtAuthGuard } from "@auth/jwt-auth.guard";
+import { RolesGuard } from "@auth/roles.guard";
+import { Read } from "@auth/roles.decorator";
+import { AuthenticatedRequest } from "@dto/internal/authenticated-request";
+import { UnauthorizedException } from "@nestjs/common";
+import {
+    checkIfUserHasReadAccessToApplication,
+    checkIfUserHasWriteAccessToApplication,
+} from "@helpers/security-helper";
 
 @ApiTags("Data Target")
 @Controller("data-target")
+@UseGuards(JwtAuthGuard, RolesGuard)
+@ApiBearerAuth()
+@Read()
+@ApiForbiddenResponse()
+@ApiUnauthorizedResponse()
 export class DataTargetController {
     constructor(private dataTargetService: DataTargetService) {}
 
     @Get()
     @ApiOperation({ summary: "Find all DataTargets" })
     async findAll(
+        @Req() req: AuthenticatedRequest,
         @Query() query?: ListAllDataTargetsDto
     ): Promise<ListAllDataTargetsReponseDto> {
-        return await this.dataTargetService.findAndCountAllWithPagination(
-            query
-        );
+        if (req.user.permissions.isGlobalAdmin) {
+            return await this.dataTargetService.findAndCountAllWithPagination(
+                query
+            );
+        } else {
+            const allowed = req.user.permissions.getAllApplicationsWithAtLeastRead();
+            if (
+                query.applicationId &&
+                !allowed.some(x => x === query.applicationId)
+            ) {
+                throw new UnauthorizedException();
+            }
+
+            return await this.dataTargetService.findAndCountAllWithPagination(
+                query,
+                allowed
+            );
+        }
     }
 
     @Get(":id")
     @ApiOperation({ summary: "Find DataTarget by id" })
-    async findOne(@Param("id") id: number): Promise<DataTarget> {
+    async findOne(
+        @Req() req: AuthenticatedRequest,
+        @Param("id") id: number
+    ): Promise<DataTarget> {
         try {
-            return await this.dataTargetService.findOne(id);
+            const dataTarget = await this.dataTargetService.findOne(id);
+            checkIfUserHasReadAccessToApplication(
+                req,
+                dataTarget.application.id
+            );
+            return dataTarget;
         } catch (err) {
             throw new NotFoundException(ErrorCodes.IdDoesNotExists);
         }
@@ -49,8 +96,13 @@ export class DataTargetController {
     @ApiOperation({ summary: "Create a new DataTargets" })
     @ApiBadRequestResponse()
     async create(
+        @Req() req: AuthenticatedRequest,
         @Body() createDataTargetDto: CreateDataTargetDto
     ): Promise<DataTarget> {
+        checkIfUserHasWriteAccessToApplication(
+            req,
+            createDataTargetDto.applicationId
+        );
         return await this.dataTargetService.create(createDataTargetDto);
     }
 
@@ -59,18 +111,34 @@ export class DataTargetController {
     @ApiOperation({ summary: "Update an existing DataTarget" })
     @ApiBadRequestResponse()
     async update(
+        @Req() req: AuthenticatedRequest,
         @Param("id") id: number,
         @Body() updateDto: UpdateDataTargetDto
     ): Promise<DataTarget> {
-        const application = await this.dataTargetService.update(id, updateDto);
+        const oldDataTarget = await this.dataTargetService.findOne(id);
+        checkIfUserHasWriteAccessToApplication(
+            req,
+            oldDataTarget.application.id
+        );
+        if (oldDataTarget.application.id != updateDto.applicationId) {
+            checkIfUserHasWriteAccessToApplication(
+                req,
+                updateDto.applicationId
+            );
+        }
 
-        return application;
+        const dataTarget = await this.dataTargetService.update(id, updateDto);
+        return dataTarget;
     }
 
     @Delete(":id")
     @ApiOperation({ summary: "Delete an existing IoT-Device" })
     @ApiBadRequestResponse()
-    async delete(@Param("id") id: number): Promise<DeleteResponseDto> {
+    async delete(
+        @Req() req: AuthenticatedRequest,
+        @Param("id") id: number
+    ): Promise<DeleteResponseDto> {
+        checkIfUserHasWriteAccessToApplication(req, id);
         try {
             const result = await this.dataTargetService.delete(id);
